+++ aerosense/__init__.py
@@ -13,8 +13,8 @@ Usage:
 from typing import List
 
 # --- Versioning ---
-__version__: str = "v5.12.0"
-__release__: str = "2026-03-26"
+__version__: str = "v5.13.0"
+__release__: str = "2026-03-27"
 
 # --- Component Exports ---
 from .core.controller import Controller